LPS-46205 SF tests
